Working through cmd enhancements

+++ ops.sh
@@ -1,4 +1,5 @@
-#!/bin/bash
+#!/usr/bin/env bash
+
 shopt -s extglob
 
 # Determine OS
@@ -89,26 +90,35 @@ get-version() {
 
 # Main Commands
 
-_ops-composer() {
+ops-composer() {
+    cmd-doc "Run local composer. Fallback to composer in a container"
+
     if [[ -e "$(which composer)" ]]; then
         composer "$@"
         return;
     fi
 
+    if [[ $OS != 'linux' ]]; then
+        echo 'composer required. Please install.'
+        exit 1
+    fi
+
     mkdir -p "$OPS_HOME/.composer"
     mkdir -p "$OPS_HOME/.ssh"
 
+    local project="$(ops project name)"
+
     ops docker run \
-        --rm -iP \
-        -v "$(pwd):/usr/src/app" \
-        -v "$OPS_HOME/.composer:/composer" \
-        -v "$OPS_HOME/.ssh:/var/www/.ssh" \
-        -v "$SSH_AUTH_SOCK:/ssh-agent" \
+        --rm -itP \
+        -v "$(pwd):/var/www/html/$project" \
+        -v "$HOME/.composer:/var/www/.composer" \
+        -v "$HOME/.ssh:/var/www/.ssh" \
+        -v "$ssh_agent:/ssh-agent" \
         -e "SSH_AUTH_SOCK=/ssh-agent" \
-        -e "COMPOSER_HOME=/composer" \
-        -w "/usr/src/app" \
-        --label=ops.project="$(ops project name)" \
-        --user "www-data:www-data" \
+        -e "COMPOSER_HOME=/var/www/.composer" \
+        -w "/var/www/html/$project" \
+        --label=ops.project="$project" \
+        --user "$OPS_DOCKER_UID:$OPS_DOCKER_GID" \
         $OPS_DOCKER_COMPOSER_IMAGE \
         composer -n "$@"
 }
@@ -118,6 +128,8 @@ _ops-docker() {
 }
 
 ops-exec() {
+    cmd-doc "Execute a non-TTY command in a container"
+
     local service=$1
     shift
 
@@ -128,18 +140,20 @@ ops-exec() {
     ops docker exec -i $id "$@"
 }
 
-ops-help() {
-    cmd-help ops ops
-
+ops-help--after() {
     echo $(ops-version)
     echo
 }
 
 ops-logs() {
+    cmd-doc "Follow logs"
+
     system-docker-compose logs -f --tail="30" "$@"
 }
 
 ops-env() {
+    cmd-doc "Set or get a variable in project .env file"
+
     #
     # list: env
     # get:  env [key]
@@ -180,6 +194,8 @@ _ops-mc() {
 }
 
 ops-mariadb() {
+    cmd-doc "MariaDB-specific commands"
+
     cmd-run mariadb "$@"
 }
 
@@ -237,20 +253,38 @@ _ops-node() {
         "$@"
 }
 
-_ops-npm() {
+ops-npm() {
+    cmd-doc "Run local npm. Fallback to npm in a container"
+
+    if [[ -e "$(which npm)" ]]; then
+        npm "$@"
+        return;
+    fi
+
+    if [[ $OS != 'linux' ]]; then
+        echo 'npm required. Please install.'
+        exit 1
+    fi
+
+    local project="$(ops project name)"
+
     ops docker run \
         --rm -itP --init \
-        -v "$(pwd):/usr/src/app" \
-        -w "/usr/src/app" \
-        --label=ops.project="$(ops project id)" \
-        --label=traefik.enable=true \
-        --user "node" \
+        -v "$(pwd):/var/www/html/$project" \
+        -v "$HOME/.composer:/var/www/.composer" \
+        -v "$HOME/.ssh:/var/www/.ssh" \
+        -v "$ssh_agent:/ssh-agent" \
+        -e "SSH_AUTH_SOCK=/ssh-agent" \
+        -e "COMPOSER_HOME=/var/www/.composer" \
+        -w "/var/www/html/$project" \
+        --label=ops.project="$project" \
+        --user "$OPS_DOCKER_UID:$OPS_DOCKER_GID" \
         --entrypoint "npm" \
         ops-node:$OPS_VERSION \
         "$@"
 }
 
-ops-package() {
+_ops-package() {
     OPS_PROJECT_BUILD="composer laravel-mix"
 
     local project_name=$(project-name)
@@ -270,7 +304,6 @@ ops-package() {
     echo "ARG OPS_PROJECT_IMAGE${NL}ARG OPS_PROJECT_DOCROOT${NL}$prebackend${NL}$(cat $OPS_HOME/build/Dockerfile)${NL}$postbackend" \
         > $project_path/ops-Dockerfile
 
-
     docker build -f ops-Dockerfile -t $project_name:latest --no-cache \
         --build-arg OPS_PROJECT_IMAGE="imarcagency/ops-$OPS_PROJECT_BACKEND:$OPS_VERSION" \
         --build-arg OPS_PROJECT_DOCROOT="$OPS_PROJECT_DOCROOT" \
@@ -278,17 +311,9 @@ ops-package() {
 }
 
 ops-ps() {
+    cmd-doc "Show process list for ops containers"
+
     system-docker-compose ps "$@"
-}
-
-
-ops-mariadb() {
-    cmd-run mariadb "$@"
-}
-
-mariadb-help() {
-    cmd-help "ops mariadb" mariadb
-    echo
 }
 
 psql-cli() {
@@ -334,6 +359,8 @@ psql-help() {
 }
 
 ops-psql() {
+    cmd-doc "PostreSQL-specific commands"
+
     cmd-run psql "$@"
 }
 
@@ -366,20 +393,27 @@ _ops-gulp() {
 }
 
 ops-redis() {
+    cmd-doc "Run interactive redis cli"
+
     system-shell-exec redis redis-cli "$@"
 }
 
 ops-restart() {
+    cmd-doc "Restart all running containers"
     ops-stop
     ops-start
 }
 
 ops-shell() {
+    cmd-doc "Enter shell or execute command"
+
     local id=$(system-docker-compose ps -q $OPS_SHELL_BACKEND)
     local project=$(project-name)
     local command="$OPS_SHELL_COMMAND"
 
-    [[ -z $id ]] && exit
+    if [[ -z $id ]]; then
+        exit
+    fi
 
     if [[ ! -z "$1" ]]; then
         command="$@"
@@ -389,6 +423,8 @@ ops-shell() {
 }
 
 ops-link() {
+    cmd-doc "Link and start project-specific containers"
+
     local project_name=$(project-name)
 
     if [[ -z $project_name ]]; then
@@ -402,6 +438,8 @@ ops-link() {
 }
 
 ops-unlink() {
+    cmd-doc "Unlink and stop project-specific containers"
+
     local project_name=$(project-name)
 
     if [[ -z $project_name ]]; then
@@ -415,16 +453,26 @@ ops-unlink() {
 }
 
 ops-project() {
+    cmd-doc "Project-specific commands"
+
     cmd-run project "$@"
 }
 
 ops-stats() {
+    cmd-doc "Watch service stats"
+
     local ids=$(system-docker-compose ps -q)
-    [[ -z $ids ]] && exit
+
+    if [[ -z $ids ]]; then
+        exit
+    fi
+
     ops docker stats $ids
 }
 
 ops-start() {
+    cmd-doc "Start services"
+
     echo 'Starting ops services...'
     echo
 
@@ -454,6 +502,8 @@ ops-start() {
 }
 
 ops-stop() {
+    cmd-doc "Stop services"
+
     system-stop
 
     local info=$(ops docker ps -a --format '{{.ID}} {{.Label "ops.project"}}' --filter="label=ops.project")
@@ -481,6 +531,8 @@ ops-stop() {
 }
 
 ops-sync() {
+    cmd-doc "Sync remote databases/files to local project"
+
     # Ops sync assumes the following:
     #
     # - SSH access is enabled to the remote web and/or DB servers
@@ -506,16 +558,17 @@ ops-sync() {
     cd "$OPS_SITES_DIR/$OPS_PROJECT_NAME"
         #source ".env"
 
-    # best debugging helper
-    # ( set -o posix ; set ) | grep -E '^OPS_'
-
     local ssh_host="$([[ ! -z $OPS_PROJECT_REMOTE_USER ]] && echo "$OPS_PROJECT_REMOTE_USER@")"
     local ssh_host="$ssh_host$OPS_PROJECT_REMOTE_HOST"
 
-    echo $OPS_PROJECT_DB_NAME
-    echo $OPS_PROJECT_DB_TYPE
-    echo $OPS_PROJECT_REMOTE_DB_TYPE
-    echo $OPS_PROJECT_REMOTE_DB_NAME
+    if [[ ! -z "$OPS_DEBUG" ]]; then
+        # print out all OPS_ vars
+        echo
+        echo '=== START DEBUG ==='
+        ( set -o posix ; set ) | grep -E '^OPS_'
+        echo '=== END DEBUG ==='
+        echo
+    fi
 
     # sync database
     if \
@@ -617,10 +670,14 @@ _ops-jq() {
 }
 
 ops-system() {
+    cmd-doc "System-specific commands"
+
     cmd-run system "$@"
 }
 
 ops-version() {
+    cmd-doc "Show ops version"
+
     echo "ops version $OPS_VERSION"
 }
 
@@ -636,6 +693,16 @@ _ops-yarn() {
         "$@"
 }
 
+my-command() {
+    cmd-doc "Test this command"
+
+    cmd-arg verbose "-v|--verbose" "enable verbosity"
+
+    eval $(cmd-get-input "$@")
+
+    echo $verbose
+}
+
 # Site sub sommands
 
 project-docker-compose() {
@@ -648,9 +715,18 @@ project-docker-compose() {
 }
 
 project-ls() {
+    cmd-doc "List all projects in OPS_SITES_DIR"
+
+    # cmd-opt "--help" "show help"
+    cmd-arg "-v|--verbose" "enable verbose mode"
+
+    #cmd-opt "--help"
+
+    local args=$(cmd-get-flag )
+
     (
         cd $OPS_SITES_DIR
-        ls -d -1 */ | sed 's/\/$//'
+        ls -d -1 */ 2>/dev/null | sed 's/\/$//'
     )
 }
 
@@ -693,13 +769,15 @@ project-exec() {
 
     local id=$(project-docker-compose ps -q $service)
 
-    [[ -z $id ]] && exit
+    if [[ -z $id ]]; then
+        exit
+    fi
 
     ops docker exec -i $id "$@"
 }
 
 project-help() {
-    cmd-help "ops project" project
+    cmd-help "ops project" project "$@"
     echo
 }
 
@@ -707,14 +785,19 @@ project-shell-exec() {
     local id=$(project-docker-compose ps -q $1)
     shift
 
-    [[ -z $id ]] && exit
+    if [[ -z $id ]]; then
+        exit
+    fi
 
     ops docker exec -it $id "$@"
 }
 
 project-stats() {
     local ids=$(project-docker-compose ps -q)
-    [[ -z $ids ]] && exit
+
+    if [[ -z $id ]]; then
+        exit
+    fi
 
     ops docker stats $ids
 }
@@ -723,10 +806,14 @@ project-stats() {
 
 system-docker-compose() {
     COMPOSE_FILE="$OPS_HOME/docker-compose/system/base.yml"
-    COMPOSE_FILE="$COMPOSE_FILE:$OPS_HOME/docker-compose/services/traefik.yml"
+    #COMPOSE_FILE="$COMPOSE_FILE:$OPS_HOME/docker-compose/services/traefik.yml"
 
     for service in $OPS_SERVICES; do
         COMPOSE_FILE="$COMPOSE_FILE:$OPS_HOME/docker-compose/services/$service.yml"
+    done
+
+    for backend in $OPS_BACKENDS; do
+        COMPOSE_FILE="$COMPOSE_FILE:$OPS_HOME/docker-compose/backends/$backend.yml"
     done
 
     if [[ ! -z "$OPS_PUBLIC" ]]; then
@@ -735,9 +822,6 @@ system-docker-compose() {
 	    COMPOSE_FILE="$COMPOSE_FILE:$OPS_HOME/docker-compose/system/private.yml"
     fi
 
-    for backend in $OPS_BACKENDS; do
-        COMPOSE_FILE="$COMPOSE_FILE:$OPS_HOME/docker-compose/backends/$backend.yml"
-    done
 
     COMPOSE_PROJECT_NAME="ops" \
     COMPOSE_FILE=$COMPOSE_FILE \
@@ -826,7 +910,7 @@ system-config() {
 
 system-install() {
     if [[ ! -d $OPS_HOME ]]; then
-        cp -rp $OPS_SCRIPT_DIR/home $OPS_HOME
+        cp -R $OPS_SCRIPT_DIR/home $OPS_HOME
 
         source $OPS_HOME/config
 
@@ -859,9 +943,10 @@ system-install() {
 
     if [[ ! -f "$OPS_HOME/certs/self-signed-cert.key" ]]; then
         system-refresh-certs
+        exit
     fi
 
-    system-refresh-services
+    #system-refresh-services
 }
 
 system-install-mkcert() {
@@ -881,6 +966,8 @@ system-install-mkcert() {
         MKCERT_URL="https://github.com/FiloSottile/mkcert/releases/download/v$OPS_MKCERT_VERSION/mkcert-v$OPS_MKCERT_VERSION-darwin-amd64"
     fi
 
+    echo $MKCERT_URL
+
     echo "Downloading mkcert v$OPS_MKCERT_VERSION"
     curl -L --silent --output $OPS_HOME/bin/mkcert-$OPS_MKCERT_VERSION $MKCERT_URL
     chmod 744 $OPS_HOME/bin/mkcert-$OPS_MKCERT_VERSION
@@ -888,6 +975,7 @@ system-install-mkcert() {
 
 system-refresh-certs() {
     sudo --non-interactive echo 2> /dev/null
+
     if [[ $? == 1 ]]; then
         echo
         echo 'Installing self-signed certs for valid HTTPS support.'
@@ -902,9 +990,9 @@ system-refresh-certs() {
     fi
 
     local project_domains=""
-    local project_count=$(ops project ls | wc -l)
+    local project_count=$(project-ls | wc -l)
 
-    for project in $(ops project ls); do
+    for project in $(project-ls); do
         project_domains+=" *.$project.$OPS_DOMAIN"
     done
 
@@ -924,11 +1012,11 @@ system-refresh-certs() {
 
         local domain_count=$(expr $project_count + 3)
 
-        echo $domain_count
-
         mv "localhost+$domain_count-key.pem" self-signed-cert.key
         mv "localhost+$domain_count.pem" self-signed-cert.crt
     )
+
+
 }
 
 system-refresh-config() {
@@ -978,18 +1066,18 @@ system-stop() {
 }
 
 system-help() {
-    cmd-help "ops system" system
+    cmd-help "ops system" system "$@"
     echo
 }
 
 main() {
-    docker ps > /dev/null
-
-    if [[ $? != 0 ]]; then
-        exit
-    fi
-
     if [[ "$@" != "system install" ]]; then
+        docker ps > /dev/null
+
+        if [[ $? != 0 ]]; then
+            exit
+        fi
+
         validate-config
     fi
 
@@ -1017,9 +1105,10 @@ fi
 # options that can be overridden by global config
 
 declare -x OPS_ENV="dev"
-declare -x OPS_BACKENDS="$OPS_BACKENDS apache-php73"
-declare -x OPS_SERVICES="$OPS_SERVICES dnsmasq portainer nginx dashboard mariadb postgres redis adminer"
-declare -x OPS_DOCKER_COMPOSER_IMAGE=${OPS_DOCKER_COMPOSER_IMAGE-"imarcagency/ops-php71:latest"}
+declare -x OPS_DEBUG="${OPS_DEBUG}"
+declare -x OPS_BACKENDS=${OPS_BACKENDS-"apache-php73"}
+declare -x OPS_SERVICES=${OPS_SERVICES-"dnsmasq portainer dashboard mariadb postgres redis adminer"}
+declare -x OPS_DOCKER_COMPOSER_IMAGE=${OPS_DOCKER_COMPOSER_IMAGE-"imarcagency/ops-php73:latest"}
 declare -x OPS_DOCKER_NODE_IMAGE=${OPS_DOCKER_NODE_IMAGE-"imarcagency/ops-node:$OPS_VERSION"}
 declare -x OPS_DOCKER_UTILS_IMAGE=${OPS_DOCKER_UTILS_IMAGE-"imarcagency/ops-utils:$OPS_VERSION"}
 declare -x OPS_DOCKER_GID=${OPS_DOCKER_GID-""}
@@ -1034,10 +1123,16 @@ declare -x OPS_ACME_EMAIL=${OPS_ACME_EMAIL-""}
 declare -x OPS_ACME_DNS_PROVIDER=${OPS_ACME_DNS_PROVIDER-""}
 declare -x OPS_ACME_PRODUCTION=${OPS_ACME_PRODUCTION-"0"}
 declare -x OPS_ADMIN_AUTH=${OPS_ADMIN_AUTH-""}
+declare -x OPS_ADMIN_AUTH_LABEL_PREFIX=""
+
 declare -x OPS_DEFAULT_BACKEND=${OPS_DEFAULT_BACKEND-"apache-php73"}
 declare -x OPS_DEFAULT_DOCROOT=${OPS_DEFAULT_DOCROOT-"public"}
 declare -x OPS_DASHBOARD_URL="https://ops.${OPS_DOMAIN}"
 declare -x OPS_MKCERT_VERSION="1.3.0"
+
+if [[ ! $OPS_ADMIN_AUTH ]]; then
+    OPS_ADMIN_AUTH_LABEL_PREFIX="disabled-"
+fi
 
 OPS_ACME_CA_SERVER="https://acme-staging-v02.api.letsencrypt.org/directory"
 if [[ $OPS_ACME_PRODUCTION == 1 ]]; then
